+++ tests/framework/db/ActiveRecordTest.php
@@ -8,15 +8,17 @@
 namespace yiiunit\framework\db;
 
 use yii\db\ActiveQuery;
+use yii\helpers\ArrayHelper;
 use yiiunit\data\ar\ActiveRecord;
 use yiiunit\data\ar\Animal;
 use yiiunit\data\ar\BitValues;
 use yiiunit\data\ar\Cat;
 use yiiunit\data\ar\Category;
 use yiiunit\data\ar\Customer;
+use yiiunit\data\ar\Document;
+use yiiunit\data\ar\Dossier;
 use yiiunit\data\ar\CustomerQuery;
 use yiiunit\data\ar\CustomerWithConstructor;
-use yiiunit\data\ar\Document;
 use yiiunit\data\ar\Dog;
 use yiiunit\data\ar\Item;
 use yiiunit\data\ar\NullValues;
@@ -960,6 +962,97 @@ abstract class ActiveRecordTest extends DatabaseTestCase
 
         Order::$tableName = null;
         OrderItem::$tableName = null;
+    }
+
+    public function testOutdatedRelationsAreResetForNewRecords()
+    {
+        $orderItem = new OrderItem();
+        $orderItem->order_id = 1;
+        $orderItem->item_id = 3;
+        $this->assertEquals(1, $orderItem->order->id);
+        $this->assertEquals(3, $orderItem->item->id);
+
+        // Test `__set()`.
+        $orderItem->order_id = 2;
+        $orderItem->item_id = 1;
+        $this->assertEquals(2, $orderItem->order->id);
+        $this->assertEquals(1, $orderItem->item->id);
+
+        // Test `setAttribute()`.
+        $orderItem->setAttribute('order_id', 2);
+        $orderItem->setAttribute('item_id', 2);
+        $this->assertEquals(2, $orderItem->order->id);
+        $this->assertEquals(2, $orderItem->item->id);
+    }
+
+    public function testOutdatedRelationsAreResetForExistingRecords()
+    {
+        $orderItem = OrderItem::findOne(1);
+        $this->assertEquals(1, $orderItem->order->id);
+        $this->assertEquals(1, $orderItem->item->id);
+
+        // Test `__set()`.
+        $orderItem->order_id = 2;
+        $orderItem->item_id = 1;
+        $this->assertEquals(2, $orderItem->order->id);
+        $this->assertEquals(1, $orderItem->item->id);
+
+        // Test `setAttribute()`.
+        $orderItem->setAttribute('order_id', 3);
+        $orderItem->setAttribute('item_id', 1);
+        $this->assertEquals(3, $orderItem->order->id);
+        $this->assertEquals(1, $orderItem->item->id);
+    }
+
+    public function testOutdatedCompositeKeyRelationsAreReset()
+    {
+        $dossier = Dossier::findOne([
+            'department_id' => 1,
+            'employee_id' => 1,
+        ]);
+        $this->assertEquals('John Doe', $dossier->employee->fullName);
+
+        $dossier->department_id = 2;
+        $this->assertEquals('Ann Smith', $dossier->employee->fullName);
+
+        $dossier->employee_id = 2;
+        $this->assertEquals('Will Smith', $dossier->employee->fullName);
+
+        unset($dossier->employee_id);
+        $this->assertNull($dossier->employee);
+
+        $dossier = new Dossier();
+        $this->assertNull($dossier->employee);
+
+        $dossier->employee_id = 1;
+        $dossier->department_id = 2;
+        $this->assertEquals('Ann Smith', $dossier->employee->fullName);
+
+        $dossier->employee_id = 2;
+        $this->assertEquals('Will Smith', $dossier->employee->fullName);
+    }
+
+    public function testOutdatedViaTableRelationsAreReset()
+    {
+        $order = Order::findOne(1);
+        $orderItemIds = ArrayHelper::getColumn($order->items, 'id');
+        sort($orderItemIds);
+        $this->assertSame([1, 2], $orderItemIds);
+
+        $order->id = 2;
+        sort($orderItemIds);
+        $orderItemIds = ArrayHelper::getColumn($order->items, 'id');
+        $this->assertSame([3, 4, 5], $orderItemIds);
+
+        unset($order->id);
+        $this->assertSame([], $order->items);
+
+        $order = new Order();
+        $this->assertSame([], $order->items);
+
+        $order->id = 3;
+        $orderItemIds = ArrayHelper::getColumn($order->items, 'id');
+        $this->assertSame([2], $orderItemIds);
     }
 
     public function testAlias()